Namespace headers and install package
