+++ lwlog/CMakeLists.txt
@@ -1,0 +1,11 @@
+file(GLOB_RECURSE headers RELATIVE ${CMAKE_CURRENT_LIST_DIR} CONFIGURE_DEPENDS *.h)
+file(GLOB_RECURSE sources RELATIVE ${CMAKE_CURRENT_LIST_DIR} CONFIGURE_DEPENDS *.cpp)
+
+add_library(lwlog ${sources})
+add_library(lwlog::lwlog ALIAS lwlog)
+target_include_directories(lwlog
+  PUBLIC
+    $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
+    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}>)
+target_link_libraries(lwlog PUBLIC Threads::Threads)
+target_compile_features(lwlog PUBLIC cxx_std_17)
