+++ lwlog/details/pattern/alignment_formatter.h
