+++ lwlog/details/pattern/attribute.h
