+++ lwlog/details/pattern/color_format_data.h
