+++ lwlog/details/pattern/flag_data.h
