+++ lwlog/details/pattern/formatter.h
