+++ lwlog/details/stream_writer.h
