+++ lwlog/details/windows_lightweight.h
