+++ lwlog/fwd.h
