+++ lwlog/interface/sink_interface.h
