+++ lwlog/logger.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "policy/log_policy.h"
-#include "policy/sink_storage_policy.h"
-#include "interface/logger_interface.h"
+#include "lwlog/policy/log_policy.h"
+#include "lwlog/policy/sink_storage_policy.h"
+#include "lwlog/interface/logger_interface.h"
 
 namespace lwlog
 {
@@ -50,4 +50,4 @@ namespace lwlog
 	};
 }
 
-#include "logger_impl.h"
+#include "lwlog/logger_impl.h"
