+++ lwlog/policy/sink_color_policy.h
