+++ lwlog/registry.h
@@ -4,8 +4,8 @@
 #include <memory>
 #include <unordered_map>
 
-#include "fwd.h"
-#include "interface/logger_interface.h"
+#include "lwlog/fwd.h"
+#include "lwlog/interface/logger_interface.h"
 
 namespace lwlog
 {
