+++ lwlog/sinks/sink.h
@@ -1,10 +1,10 @@
 #pragma once
 
-#include "level.h"
-#include "details/pattern/pattern.h"
-#include "policy/sink_color_policy.h"
-#include "policy/threading_policy.h"
-#include "interface/sink_interface.h"
+#include "lwlog/level.h"
+#include "lwlog/details/pattern/pattern.h"
+#include "lwlog/policy/sink_color_policy.h"
+#include "lwlog/policy/threading_policy.h"
+#include "lwlog/interface/sink_interface.h"
 
 namespace lwlog::sinks
 {
@@ -33,4 +33,4 @@ namespace lwlog::sinks
 	};
 }
 
-#include "sink_impl.h"
+#include "lwlog/sinks/sink_impl.h"
