+++ lwlog/sinks/stderr_sink.h
@@ -1,28 +1,28 @@
 #pragma once
 
-#include "sink.h"
-#include "details/stream_writer.h"
+#include "lwlog/sinks/sink.h"
+#include "lwlog/details/stream_writer.h"
 
 namespace lwlog::sinks
 {
 	template<typename ThreadingPolicy>
-	class stdout_sink 
+	class stderr_sink 
 		: public sink<colored_policy, ThreadingPolicy>
 		, public details::stream_writer
 	{
 	  using sink_t = sink<colored_policy, ThreadingPolicy>;
 	public:
-		stdout_sink();
+		stderr_sink();
 		void sink_it(const details::log_message& log_msg) override;
 	};
 
 	template<typename ThreadingPolicy>
-	stdout_sink<ThreadingPolicy>::stdout_sink()
-		: details::stream_writer(stdout)
+	stderr_sink<ThreadingPolicy>::stderr_sink()
+		: details::stream_writer(stderr)
 	{}
 
 	template<typename ThreadingPolicy>
-	void stdout_sink<ThreadingPolicy>::sink_it(const details::log_message& log_msg)
+	void stderr_sink<ThreadingPolicy>::sink_it(const details::log_message& log_msg)
 	{
 		details::stream_writer::write(sink_t::m_pattern.compile(log_msg));
 	}
